fix(hw): incorrect PPTC footprint
This commit fixes the incorrect PPTC footprint size.

The existing schematic specifies the Bourns MF-MSMF030-2 part which
should be using the `Fuse_1812_4532Metric` footprint but was instead
assigned the `Fuse_1206_3216Metric` footprint. As a result, the
component CANNOT be placed on the circuit board.

To rectify this, we changed the component to use the Littelfuse
1206L035/16 part instead, which is almost similar to the specifications
of the original part.

Importantly, a deliberate decision was made to swap component instead of
correcting the footprint in order to avoid the significant change to the
board layout required to accommodate the correct but larger 1812
footprint.